Change reject-data to reject-site-visit
