+++ bims/api_urls.py
@@ -34,10 +34,11 @@ from bims.api_views.reference_list import ReferenceList, ReferenceEntryList
 from bims.api_views.search import CollectionSearchAPIView
 from bims.api_views.thermal_data import ThermalDataApiView
 from bims.api_views.validate_object import ValidateSite
-from bims.api_views.reject_object import RejectSite, RejectTaxon
+from bims.api_views.reject_object import (
+    RejectSite, RejectTaxon, RejectSiteVisit
+)
 from bims.api_views.taxon_images import TaxonImageList
 from bims.api_views.validate_object import ValidateObject
-from bims.api_views.reject_data import RejectData
 from bims.api_views.get_biorecord import (
     GetBioRecordDetail,
     GetBioRecords,
@@ -162,8 +163,8 @@ urlpatterns = [
         name='list-non-biodiversity-layer'),
     url(r'^validate-object/$',
         ValidateObject.as_view(), name='validate-object'),
-    url(r'^reject-data/$',
-        RejectData.as_view(), name='reject-data'),
+    url(r'^reject-site-vist/$',
+        RejectSiteVisit.as_view(), name='reject-site-visit'),
     url(r'^get-bio-object/$',
         GetBioRecordDetail.as_view(), name='get-bio-object'),
     url(r'^get-site-code/$',
