+++ bims/api_views/reject_object.py
@@ -1,6 +1,5 @@
 # coding=utf-8
 from bims.models.survey import Survey
-from django.http.response import Http404
 
 from bims.models.taxonomy import Taxonomy
 
@@ -40,6 +39,35 @@ class RejectCollectionData(UserPassesTestMixin, LoginRequiredMixin, APIView):
             )
             return JsonResponse({'status': 'success'})
         except BiologicalCollectionRecord.DoesNotExist:
+            return HttpResponse(
+                'Object Does Not Exist',
+                status=status.HTTP_400_BAD_REQUEST
+            )
+
+
+class RejectSiteVisit(UserPassesTestMixin, LoginRequiredMixin, APIView):
+
+    def test_func(self):
+        return self.request.user.has_perm('bims.can_validate_site_visit')
+
+    def handle_no_permission(self):
+        messages.error(self.request, 'You don\'t have permission '
+                                     'to reject site vist')
+        return super(RejectSiteVisit, self).handle_no_permission()
+
+    def get(self, request):
+        object_pk = request.GET.get('pk', None)
+        rejection_message = request.GET.get('rejection_message', None)
+        try:
+            survey = Survey.objects.get(pk=object_pk)
+            survey.ready_for_validation = False
+            survey.reject(
+                rejection_message=rejection_message
+            )
+            survey.save()
+
+            return JsonResponse({'status': 'success'})
+        except Survey.DoesNotExist:
             return HttpResponse(
                 'Object Does Not Exist',
                 status=status.HTTP_400_BAD_REQUEST
